Bootstrap RabbitMQ configuration to separate module
