+++ pet-hub-rabbitmq/src/main/java/com/pet/pethubrabbitmq/util/RabbitMqListenerLogger.java
@@ -1,4 +1,4 @@
-package com.pet.pethubbatch.infrastructure.logging;
+package com.pet.pethubrabbitmq.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.Message;
